requirements.txt
langchain 
langchain_ollama
selenium
html5lib
python-dotenv
streamlit 
beautifulsoup4
lxml